dt/overlays: Add PiFace Digital Device Tree Overlay
The PiFace Digital is a convenient breakout board for the Microchip
mcp23s17 SPI GPIO port expander.

The first eight GPIOs 0..7 (bank A) are connected to eight output
terminals and LEDs, plus two relays on the first two outputs. These
output loads are active-high.

The next eight GPIOs 8..15 (bank B) are connected to eight input
terminals with four on-board switches connecting them to ground. Inputs
devices are therefore expected to bridge terminals to ground, so the
mcp23s17 pullups are activated for GPIO bank B.

Signed-off-by: Thomas Preston <thomas.preston@codethink.co.uk>